refactor: Notification marking modularization
- remove `emitRoomEvent` integration in `UserNotifications.jsx`.
- remove the notification emission handling for `mark as read` and `mark all as read` events from `UserNotifications.jsx`.
- create new handler functions `emitMarkAsRead` and `emitMarkAllAsRead` to separate concerns and modularize logic in `notificationHandlers.js`.
- integrate `emitMarkAsRead` and `emitMarkAllAsRead` in the respective `onClick` handlers in `UserNotifications.jsx`.
